fix the row count refresh issue
